Configurable option allowing Session.create to work from any route
Currently if you try to use `Session.create` outside of the `/auth`
routes it will fail with the error `Cannot use Auth.privateKey.` This is
due to the privateKey binding only on the `/auth` route.

In some scenarios you may with want to call
[Session](https://docs.sst.dev/clients/auth#session) methods in routes
other than `/auth`. This commit updates the Auth stacks `attach` method
to allow this. An additional optional attribute `routes` is added to the
to `ApiAttachmentProps`. Any additional routes specified will have the
privateKey bound.